REF: login, register, bd
